requirements.txt
# MCP Server Dependencies
mcp>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
sqlite-utils>=3.35

# Utilities
python-dotenv>=1.0.0